- Hotkeys are now dynamically mapped using api.Input.Hotkeys to support user-customized keybindings.
- Added toggleStepUpKeyHeld and reloadConfigKeyHeld state variables to prevent repeated actions when hotkeys are held down.
These states are reset using the OnKeyUp event, ensureing that actions like enabling/disabling StepUp or reloading the configuration execute only once per key press.

- Duplicate messages ("Step height cannot exceed 2.0 blocks" or "Step height cannot be less than 0.5 blocks") when holding down the increment or decrement keys are eliminated. Messages are now displayed only when explicitly attempting to go beyond the limit.

- Automatically adjusts invalid StepHeight, StepHeightIncrement and DefaultHeight values in the configuration file:
StepHeight values above the maximum (2.0) are clamped to the maximum.
StepHeightIncrement values below the minimum (0.1) are clamped to the minimum.
DefaultHeight values above the maximum or below the minimum will reset to the original default value '0.6'
Invalid values are logged to inform users of the adjustments.

- Now supports multiple versions of Vintage Story, compatible with v1.19.8 (and possibly older versions? Not tested)